nmxact - image upgrade command
This command performs the following procedure:

1. Send image erase command.
2. If the image erase command succeeded, proceed to step 5.
3. Else if the peer is disconnected, attempt to reconnect to the peer.
   If the reconnect attempt fails, abort the command and report the
   error.  If the reconnect attempt succeeded, proceed to step 5.
4. Else (the erase command failed and the peer is still connected),
   proceed to step 5.
5. Execute the upload command.

The newtmgr tool now uses the upgrade command when the user types "image
upload".

This command is necessary because a plain upload command fails on some
combinations of device and client when sent over BLE.  What actually
happens is the target device stalls while processing the first upload
request, causing the BLE connection to drop.  This happens because the
first request triggers a flash erase, and the device is running out of
flash.