newtmgr - workaround macOS ble lib mtu issue
This is a workaround for a ble library issue when running on macOS.  In
macOS, the request to exchange MTU doesn't actually do anything.  The
ble library relies on the assumption that the OS already exchanged MTUs
on its own.  If this assumption is incorrect, the number that was
returned is the default out of date value (23).  In this case, newtmgr
sleeps for one second and requeries the library.  A maximum of three queries
are attempted.