nmxact - Don't stop listening for OMP on disc.
Rather than stop the OMP listener on disconnect, stop it on session
close.  This fixes the issue that occurs when a BLE session lives on
after a disconnect, i.e., when a connection retry occurs due to MTU
exchange timeout.

There are a lot of synchronization issues here.  It is probably time for
a large round of cleanup and "refactoring" (aka rewriting).