nmxact - Send CoAP responses in Go routine.
This unblocks blehostd while nmxact is generating its CoAP response.
Prior to this change, the host is waiting for an ATT error code to send
in response to the incoming characteristic write.  Now, nmxact
immediately indicates an error code of 0, and the notification
containing the CoAP response gets sent later.