nmxact - fix for unready controller
If the controller is in a bad state when blehostd is started, the
controller won't respond to the host's first HCI command.  In turn, the
host indicates a stack reset to nmxact.  If the stack resets before
host<->controller sync is acquired, nmxact should ignore it and continue
trying to sync.

Prior to this commit, the reset event would cause the BLE transport to
abort its start operation.