nmxact - conn_cancel_evt listener uses wrong seq
When blehostd sends a `conn_cancel_evt` event, it uses the sequence
number associated with the connection, not the sequence number
corresponding to the `conn_cancel` request.

Prior to this change, nmxact would listen for the `conn_cancel_evt` with
a sequence number equal to that of the request.  The event would never
arrive, which caused the transport to be restarted.

Now, the connect listener listens for the cancel event, in addition to
the connect event.