file download - get full file, not just 1st chunk
Prior to this commit, the `file download` command was only retrieving
the first chunk of the specified file.  The command handler was
expecting the len (file length) field to be populated in every response,
but it is actually only populated in the initial response.
Consequently, the handler behaved as though the file lenght were
truncated to 0 upon receiving the second response.