res - Allow integer values
Prior to this commit, newtmgr assumed the value in a res key-value-pair
was a string.  So, the following command:

    res put public myresource key=1

would include the string "1" in the PUT request.

This commit changes this behavior such that newt encodes integers as
numbers in the CoAP request.  To force the string type, enclose the
value in double-quotes, e.g.,

    res put public myresource 'key="1"'