nmxact: Add data hash to 1st chunk of image upload data
When uploading image, 1st chunk will include "datahash" field which is
simply SHA-256 calculated over complete data to be uploaded (so it is
not the same as image hash). This field can be then used by imgmgr on
target to detect whether the same data is being uploaded and possibly
resume transfer which was interrupted previously.