Allow "conn_timeout" for ble and bll types.
This specifies the connection timeout, in seconds.  E.g.,

    newtmgr --conntype ble --connstring peer_name=mydev,conn_timeout=23.54' -t 4 echo a

This command allows 23.54 seconds for the BLE connection to be
established.  Upon sending the echo request, newtmgr waits four seconds
before reporting a response timeout.

In addition, if the `conn_timeout` setting is not specified, it defaults
to the response timeout value (`-t`).