Fix [un]marshal of log-entry-type
The code was attempting to marshal and unmarshal the log-entry-type as a
CBOR string.  That is, rather than converting to and from a raw string,
it was using CBOR encoding (type+length prefix).  This is not correct;
the raw values of the log response struct are ultimately encoded or
decoded using the CBOR codec library, so encoding an individual field as
CBOR causes a double encoding.

Also, the function type of `(l LogEntryType) MarshalBinary()` was
incorrect.  It should take no arguments and return `([]byte, error)`.