image: Multiple rounds of chunklen reduction
Prior to this commit, newtmgr determined how many bytes of image data to
include in an upload request as follows:

1. Choose a large number to start (512).
2. Encode the upload request into CoAP (+ optional security).
3. Compare the size of request to the transport's MTU. If the request is
too large, subtract the difference from the chunk length.

This commit changes the above procedure such that it runs in a loop.
After reducing the chunk length, a new request is encoded and its length
is compared to the MTU.  The loop executes until the encoded request
fits within the MTU.