res: `-j` option to specify payload as JSON
Before this commit, the `res` command accepted a payload in the form of
`k=v` pairs, e.g.,

    newtmgr res put /my/res a=start dur=1

This commit adds a new option to the `res` command: `-j`.  When this
option is specified, the CoAP message body is expressed as a JSON
string, e.g.,

    newtmgr res -j put /my/res '{"a":"start","dur":1}'

This allows for more complex payloads, and it removes ambiguity about
what data types are used.