Remove security requirements from CoAP commands
This commit removes a feature that has been unused for years.

Initially, the idea was that each CoAP resource would be defined with
one of the following security levels:

    * public (no security)
    * unauthenticated (encryption)
    * secure (encryption and authentication)

A session (connection) would be secured in a transport-specific way, but
the main transport under consideration was Bluetooth.  In Bluetooth,
characteristics are assigned security levels resembling the three above.
To implement this defunct security mechanism in CoAP-over-Bluetooth, we
would need three different characteristics, one corresponding to each
security level.  When sending a CoAP request, the newtmgr client would
choose the correct characteristic based on the secure state of the
connection.

This security policy has not been supported in Mynewt for a few years.
The code implementing it is unused and unmaintained, so I am removing it
now.