nmxact: Don't require serial session to "unlisten"
There is a max of one active session per serial transport.  This
limitation is necessary because there is no demuxing protocol.  Whatever
data is received over serial is just forwarded to the active session.

Before this commit, nmxact required the active sesion to unregister
itself before any session could be registered again, even if the two
sessions are the same.

This commit relaxes this restriction.  Now, it is OK to register the
same session multiple times without unregistering in between.

This change is necessary due to the recent change to the session API.
Now, listening and transmitting are separate operations, so it is not
trivial to arrange for a session to be unregistered after use.  In
practice, there is never more than one serial session anyway.