Include imghash in log show output
The Mynewt log entry header optionally includes an "imghash" field.
This field contains a truncated hash of the build that was running when
the entry was logged.  If an entry contains this hash, the Mynewt device
includes it in a `log show` response.

This commit changes newtmgr to print the image hash when it sees it in a
`log show` response.