nmxact: Create a listener for each OMP response
Prior to this change, nmxact created a single CoAP listener to handle
all incoming OMP responses.  This listener would parse the embedded NMP
message, then dispatch to the appropriate NMP listener based on the NMP
sequence number.

There is a problem with this approach.  If the server rejects an
outgoing OMP request with a CoAP error (e.g., the omp resource
requires a more secure connection), then the error does not get passed
up to the caller.  The error response does not contain an embedded NMP
response, so there is no way to match it with the appropriate listener.
The result is that nmxact drops the CoAP error, and the OMP operation
eventually times out.

This commit makes the following changes:
    * Every outgoing OMP request contains a single-byte token consisting
      of the NMP sequence number.
    * A new CoAP listener and NMP listener are created for each expected
      OMP response.
    * Both NMP errors and CoAP errors are communicated to the caller.

With this change, when an OMP elicits a CoAP error, newtmgr displays the
error and terminates.