README: Indicate that GO111MODULE must be set
Now that newtmgr uses modules instead of a vendor directory, module
support needs to be enabled to build newtmgr.

This setting is only necessary for older versions of go.  In newer
versions, the setting is enabled by default (outside of the GOPATH).
Maybe we can revert this change once modules become fully mainstream.