nmxact: Remove "Serialized OMP" debug log message
This message got logged immediately before the request got transmitted.
The result was two nearly identical large log messages (both contain the
full payload of the request).

In addition, the image upload command actually serializes several
requests for each one it sends.  It does this to determine the ideal
chunk size to use for the next request.  The result is even more noise
in the logs