nmxact: Don't skip seq numbers in `image upload`
nmxact generates several image upload requests for each one that it
actually sends.  It does this to determine the ideal chunk size.

Each of these requests was consuming a sequence number.  As a result,
the sequence number would increase by three for each request that was
actually transmitted.

This commit changes the image upload command such that the sequence
number is only increased when a request is actually sent.