res: Add `--int` option to parse numbers as ints
JSON only supports a single numeric type: "number".  When the `res -j`
command parses a payload, the JSON unmarshal function parses all numbers
as instances of float64.  This makes the resulting CBOR payload
incorrect when the user just wants to use integers.

This commit adds an option to the res command: `--int`.  When specified,
all numbers in the payload are converted to integers when the conversion
does not result in a change in value.  For example, `1.0` is converted
to an integer, but `1.5` is not.