log show: Specify `-a` to get entire log
Normally, the `newtmgr log show` command only receives one packet's
worth of log entries.

This commit adds a `-a` option to this command.  When this option is
specified, newtmgr continues requesting more data until it reaches the
end of the log.

This is slightly better than using a script to read an entire log
because newtmgr will reuse the established BLE connection for its
subsequent requests, making the process faster.