res: Use zero payload when no args specified
This command:

    newtmgr res get /my/res

should send a CoAP GET request with *no* payload.  An optional payload
can be specified with extra arguments after the resource name.  Since
there are no extra arguments, there should be no payload.

Prior to this commit, newtmgr was including an empty map as the payload
in such requests.

The bug was caused by this quirk in Go:
https://forum.golangbridge.org/t/a-nil-a-b-b-nil-with-pointers-and-interface/10593

In short, if `nil` is assigned to a variable of type `interface{}`, the
variable will *not* compare equal to `nil`.  An interface variable
contains two pieces of information: 1) its type, and 2) its value.  The
variable is not equal to `nil` because `nil` is an *untyped* nil value.

The solution is to return the untyped `nil` rather than the nil
interface{}.