image: optimize Device Firmware Upgrade (DFU)
Optimize the DFU process by queuing multiple chunks of the
image at once. The window of outstanding requsts is slowly
increased until errors are seen.

The process is started with a minimum window size of outstanding
requests. As the successful responses come in the size is increased.

Internal map is maintained to track if the CoAP server (device)
has missed a chunk and repeatedly requests for it in the response.

As several requests are outstanding at the same time, this
optimizes particularly Bluetooth transport.

Serial and Mtech LORA is updated for compliance with the updated
interface, although they use underlying synchronous transmit/receive
functionality.

Signed-off-by: Naveen Kaje <naveen.kaje@juul.com>