nmcoap: Replace filter callbacks with interfaces
Sometimes it is desirable to reuse the same parameters when generating a
sequence of messages (e.g., when calculating the maximum payload size
for an image upload request).  This was not possible with a session
configured with a TxFilter since there is no way to specify which
parameters the filter should use for a given message.

This commit creates a pair of interfaces (TxFilter and RxFilter) to
replaces the old callbacks.  The TxFilter interface allows its
parameters to be "frozen" and "unfrozen".