Fix CoAP over serial fragmentation problem
When sending large packets over serial newtmgr was fragmenting
them twice in case of using CoAP over serial. First fragmentation
was performed inside Transceiver's TxCoap method,
before passing data to Tx method from SerialXport module.
Tx method from SerialXport then would once again fragment
data with the same MTU, but on this stage each fragment would get
a packet stat designator, which is necessary to determine if
a specific fragment is a start of new packet or continuing one.
Because of previous data fragmentation, each fragment was passed
to SerialXport Tx method individualy and each of this fragments
was incorrectly interpreted as a starting of a new packet.

To solve this issue, the fragmentation was moved from
Transceiver's TxCoap method to txCb functions, so now every
kind of transport can implement it's own fragmentation.